+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
+## [Unreleased](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v0.8.0...HEAD)
+- Additional extension info (version and DB schema name) in `borealis-pg:extensions` and `borealis-pg:extensions:install` command output
+
 ## [0.8.0](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v0.7.0...v0.8.0)
 - Adds the `borealis-pg:psql` command to launch an interactive psql session through a secure tunnel directly from the CLI